backend/pyproject.toml
[project]
name = "idpportal-backend"
version = "0.1.0"
description = "IDP Portal AI Backend - Multi-agent platform engineering system"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.5.0",
    "langgraph>=0.2.0",
    "langchain-core>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-anthropic>=0.3.0",
    "mcp>=1.0.0",
    "httpx>=0.27.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.29.0",
    "pgvector>=0.3.0",
    "alembic>=1.13.0",
    "python-jose[cryptography]>=3.3.0",
    "python-multipart>=0.0.9",
    "pyyaml>=6.0.0",
    "prometheus-fastapi-instrumentator>=7.0.0",
    "structlog>=24.0.0",
    "tenacity>=9.0.0",
    "redis>=5.0.0",
    "sse-starlette>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "httpx",
    "ruff>=0.8",
    "coverage>=7.0",
]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
